# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handler.CancelInventoryUsageRequest:
    properties:
      reason:
        example: 临时有事
        type: string
    type: object
...
  internal_handler.UseInventoryRequest:
    properties:
      address_id:
        type: integer
      delivery_type:
        example: 1
        type: integer
      quantity:
        example: 1
        type: integer
      remark:
        type: string
    required:
    - delivery_type
    type: object
...
  yujixinjiang_backend_internal_model.Account:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      is_rider:
        type: integer
      last_login_at:
        type: string
      nickname:
        type: string
      openid:
        type: string
      phone:
        type: string
      status:
        type: integer
      type:
        type: integer
      unionid:
        type: string
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.AddressSnapshot:
    properties:
      city:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      detail:
        type: string
      district:
        type: string
      province:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.DeliveryOrder:
    properties:
      accepted_at:
        type: string
      created_at:
        type: string
      deliver_photos:
        items:
          type: string
        type: array
      deliver_remark:
        type: string
      delivered_at:
        type: string
      id:
        type: integer
      inventory_usage:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.UserInventoryUsage'
      inventory_usage_id:
        type: integer
      order:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Order'
      order_id:
        type: integer
      rider_id:
        type: integer
      started_at:
        type: string
      status:
        type: integer
      updated_at:
        type: string
      user_confirmed:
        type: integer
    type: object
  yujixinjiang_backend_internal_model.MerchantProfile:
    properties:
      account:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Account'
      account_id:
        type: integer
      address:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      shop_logo:
        type: string
      shop_name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_model.Order:
    properties:
      account_id:
        type: integer
      activity_id:
        type: integer
      address_snapshot:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.AddressSnapshot'
      created_at:
        type: string
      delivery_type:
        type: integer
      discount_amount:
        type: number
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/yujixinjiang_backend_internal_model.OrderItem'
        type: array
      merchant_id:
        type: integer
      merchant_review_stage:
        type: integer
      order_no:
        type: string
      paid_at:
        type: string
      pay_amount:
        type: number
      pay_status:
        type: integer
      remark:
        type: string
      status:
        type: integer
      total_amount:
        type: number
      updated_at:
        type: string
      user_coupon_id:
        type: integer
    type: object
  yujixinjiang_backend_internal_model.OrderItem:
    properties:
      activity_id:
        type: integer
      activity_product_id:
        type: integer
      created_at:
        type: string
      group_buy_id:
        type: integer
      group_buy_team_id:
        type: integer
      id:
        type: integer
      order_id:
        type: integer
      product_id:
        type: integer
      product_image:
        type: string
      product_name:
        type: string
      purchase_type:
        type: integer
      quantity:
        type: integer
      spec:
        type: string
      subtotal:
        type: number
      unit_price:
        type: number
    type: object
  yujixinjiang_backend_internal_model.Product:
    properties:
      category:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.ProductCategory'
      category_id:
        type: integer
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      enable_coupon:
        type: integer
      enable_group_buy:
        type: integer
      group_buy_allow_repeat:
        type: integer
      group_buy_price:
        type: number
      group_buy_target_count:
        type: integer
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_hot:
        type: integer
      item_type:
        type: integer
      merchant:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.MerchantProfile'
      merchant_id:
        type: integer
      name:
        type: string
      original_price:
        type: number
      price:
        type: number
      sales_count:
        type: integer
      status:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_model.ProductCategory:
    properties:
      created_at:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.UserInventory:
    properties:
      account_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      last_order_id:
        type: integer
      product:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      spec:
        type: string
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_model.UserInventoryUsage:
    properties:
      account_id:
        type: integer
      address_snapshot:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.AddressSnapshot'
      cancel_reason:
        type: string
      created_at:
        type: string
      delivery_order:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.DeliveryOrder'
      delivery_order_id:
        type: integer
      delivery_type:
        type: integer
      id:
        type: integer
      inventory:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.UserInventory'
      inventory_id:
        type: integer
      merchant_id:
        type: integer
      product:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      remark:
        type: string
      source_order_id:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_query.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  yujixinjiang_backend_internal_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  yujixinjiang_backend_internal_service.BuyerBrief:
    properties:
      account_id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
    type: object
...
  yujixinjiang_backend_internal_service.DeliveryView:
    properties:
      accepted_at:
        type: string
      created_at:
        type: string
      deliver_photos:
        items:
          type: string
        type: array
      deliver_remark:
        type: string
      delivered_at:
        type: string
      id:
        type: integer
      inventory_usage:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.UserInventoryUsage'
      inventory_usage_id:
        type: integer
      order:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Order'
      order_id:
        type: integer
      rider_id:
        type: integer
      started_at:
        type: string
      status:
        type: integer
      status_text:
        type: string
      updated_at:
        type: string
      user_confirmed:
        type: integer
    type: object
...
  yujixinjiang_backend_internal_service.InventoryUsageView:
    properties:
      account_id:
        type: integer
      address_snapshot:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.AddressSnapshot'
      buyer:
        $ref: '#/definitions/yujixinjiang_backend_internal_service.BuyerBrief'
      cancel_reason:
        type: string
      created_at:
        type: string
      delivery_order:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.DeliveryOrder'
      delivery_order_id:
        type: integer
      delivery_type:
        type: integer
      id:
        type: integer
      inventory:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.UserInventory'
      inventory_id:
        type: integer
      merchant_id:
        type: integer
      product:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      remark:
        type: string
      source_order_id:
        type: integer
      status:
        type: integer
      status_text:
        type: string
      updated_at:
        type: string
      verify_code:
        type: string
    type: object
...
info:
  contact:
    email: support@yujixinjiang.com
    name: API Support
  description: 豫记信疆微信小程序后端接口文档，支持在线调试。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 豫记信疆 API
  version: "1.0"
paths:
...
  /user/inventory/{id}/use:
    post:
      consumes:
      - application/json
      description: 指定数量扣减库存并创建使用记录；自提时返回核销码
      parameters:
      - description: 背包记录 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 使用方式
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UseInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.InventoryUsageView'
              type: object
      security:
      - BearerAuth: []
      summary: 使用背包商品
      tags:
      - 用户-背包
  /user/inventory/usages:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_query.PageResult'
              type: object
      security:
      - BearerAuth: []
      summary: 背包使用记录列表
      tags:
      - 用户-背包
  /user/inventory/usages/{id}:
    get:
      parameters:
      - description: 使用记录 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.InventoryUsageView'
              type: object
      security:
      - BearerAuth: []
      summary: 背包使用记录详情
      tags:
      - 用户-背包
  /user/inventory/usages/{id}/cancel:
    post:
      consumes:
      - application/json
      description: 自提/未接单配送：立即取消并回滚库存；骑手已接单至用户确认收货前：提交取消申请，商家审核通过后回滚库存
      parameters:
      - description: 使用记录 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 取消原因
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.CancelInventoryUsageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.InventoryUsageView'
              type: object
      security:
      - BearerAuth: []
      summary: 取消背包使用
      tags:
      - 用户-背包
...
  /user/orders/{id}/confirm-receipt:
    post:
      description: 购买订单配送场景，等价于确认关联配送单
      parameters:
      - description: 订单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.DeliveryView'
              type: object
      security:
      - BearerAuth: []
      summary: 按订单确认收货
      tags:
      - 用户-订单
...
securityDefinitions:
  BearerAuth:
    description: '登录后获取的 JWT，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"